impl versionning

+++ src/main/java/com/asim/curd_demo/model/userbalance/UserBalanceDTO.java
@@ -12,5 +12,5 @@ public class UserBalanceDTO {
     @JsonProperty("reverse_balance")
     private double reverseBalance;
 
-    private float version;
+    private int version;
 }
